refactor: 결과 페이지 리팩토링 - subPage 폴더를 가지는 대신 *.subPage.tsx 파일로 분리 - page, subPage가 가지고 있는 로직은 *.logic.ts 파일로 분리하여 관리 - 자세한 내용 이슈 참조
